Refactor camera controller into hook
